Add a 404 page
On the new server, there is no default 404 page, which results in an
empty document if navigating to an invalid URL. This generates a 404
page using our page CSS, so that it fits into the rest of the site
better.

This also tweaks the layout so that the footer is forced to the bottom
even for small pages (like the 404 page).